pyproject.toml
[project]
name = "lacrei-api"
version = "0.1.0"
description = ""
authors = [
    {name = "EwertonHecsley",email = "ewerton.martinscomercial@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "lacrei_api", from = "src"}]

[tool.poetry.dependencies]
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
